TE-2527 Move test and dev dependencies out of base.in
