+++ common/lib/xmodule/xmodule/partitions/partitions_service.py
@@ -13,8 +13,6 @@ from xmodule.modulestore.django import modulestore
 
 log = logging.getLogger(__name__)
 
-
-# settings will not be available when running nosetests.
 FEATURES = getattr(settings, 'FEATURES', {})
