+++ pavelib/utils/test/utils.py
@@ -10,7 +10,10 @@ from paver.easy import cmdopts, sh, task
 from pavelib.utils.envs import Env
 from pavelib.utils.timer import timed
 
-from bok_choy.browser import browser
+try:
+    from bok_choy.browser import browser
+except ImportError:
+    browser = None
 
 MONGO_PORT_NUM = int(os.environ.get('EDXAPP_TEST_MONGO_PORT', '27017'))
 MINIMUM_FIREFOX_VERSION = 28.0
