+++ requirements/edx/base.txt
@@ -13,7 +13,6 @@
 git+https://github.com/solashirai/crowdsourcehinter.git@518605f0a95190949fe77bd39158450639e2e1dc#egg=crowdsourcehinter-xblock==0.1
 git+https://github.com/mitodl/django-cas.git@afac57bc523f145ae826f4ed3d4fa8b2c86c5364#egg=django-cas==2.1.1
 git+https://github.com/edx/django-celery.git@756cb57aad765cb2b0d37372c1855b8f5f37e6b0#egg=django-celery==3.2.1+edx.2
-git+https://github.com/hmarr/django-debug-toolbar-mongo.git@b0686a76f1ce3532088c4aee6e76b9abe61cc808#egg=django-debug-toolbar-mongo==0.1.10
 git+https://github.com/edx/django-oauth-plus.git@01ec2a161dfc3465f9d35b9211ae790177418316#egg=django-oauth-plus==2.2.9.edx-1
 git+https://github.com/edx/django-openid-auth.git@0.15.1#egg=django-openid-auth==0.15.1
 git+https://github.com/jazzband/django-pipeline.git@d068a019169c9de5ee20ece041a6dea236422852#egg=django-pipeline==1.5.3
@@ -24,7 +23,6 @@ git+https://github.com/edx/django-rest-framework.git@1ceda7c086fddffd1c440cc8685
 -e git+https://github.com/edx/DoneXBlock.git@01a14f3bd80ae47dd08cdbbe2f88f3eb88d00fba#egg=done-xblock
 -e git+https://github.com/jazkarta/edx-jsme.git@690dbf75441fa91c7c4899df0b83d77f7deb5458#egg=edx-jsme
 git+https://github.com/mitodl/edx-sga.git@6b2f7aa2a18206023c8407e2c46f86d4b4c3ac96#egg=edx-sga==0.8.0
-git+https://github.com/edx/lettuce.git@31b0dfd865766243e9b563ec65fae9122edf7975#egg=lettuce==0.2.23+edx.1
 git+https://github.com/edx/xblock-lti-consumer.git@v1.1.7#egg=lti_consumer-xblock==1.1.7
 git+https://github.com/edx/MongoDBProxy.git@25b99097615bda06bd7cdfe5669ed80dc2a7fed0#egg=MongoDBProxy==0.1.0
 -e .
@@ -53,28 +51,21 @@ argparse==1.4.0
 asn1crypto==0.24.0
 attrs==17.2.0
 babel==1.3
-beautifulsoup4==4.1.3
 beautifulsoup==3.2.1      # via pynliner
-before_after==0.1.3
 billiard==3.3.0.23        # via celery
 bleach==1.4
-bok-choy==0.7.2
 boto3==1.4.8
 boto==2.39.0
 botocore==1.8.17
 celery==3.1.25
 cffi==1.11.5
 charade==1.0.3            # via pysrt
-chrono==1.0.2
-click==6.7                # via flask, user-util
-colorama==0.3.9           # via radon
+click==6.7                # via user-util
 cryptography==2.1.4
-cssselect==0.9.1
 cssutils==1.0.2           # via pynliner
 ddt==0.8.0
 decorator==4.3.0          # via dogapi, pycontracts
 defusedxml==0.4.1
-dicttoxml==1.7.4          # via moto
 django-appconf==1.0.2     # via django-statici18n
 django-babel-underscore==0.5.2
 django-babel==0.6.2       # via django-babel-underscore
@@ -85,7 +76,6 @@ django-config-models==0.1.8
 django-cors-headers==2.1.0
 django-countries==4.6.1
 django-crum==0.7.2
-django-debug-toolbar==1.8
 django-fernet-fields==0.5
 django-filter==1.0.4
 django-ipware==1.1.0
@@ -110,12 +100,11 @@ django-user-tasks==0.1.5
 django-waffle==0.12.0
 django-webpack-loader==0.4.1
 django==1.11.12
-django_nose==1.4.1
 djangorestframework-jwt==1.11.0
 djangorestframework-xml==1.3.0  # via edx-enterprise
 dm.xmlsec.binding==1.3.3  # via python-saml
 docopt==0.6.2
-docutils==0.14
+docutils==0.14            # via botocore
 dogapi==1.2.1
 edx-ace==0.1.6
 edx-analytics-data-api-client==0.14.4
@@ -141,35 +130,22 @@ edxval==0.1.13
 elasticsearch==1.9.0      # via edx-search
 enum34==1.1.6
 event-tracking==0.2.4
-extras==1.0.0             # via python-subunit, testtools
-factory_boy==2.8.1
-faker==0.8.13             # via factory-boy
-faulthandler==3.0         # via nose-faulthandler
 feedparser==5.1.3
 firebase-token-generator==1.3.2
-flaky==3.4.0
-flask==0.12.2             # via moto
-freezegun==0.3.8
 fs-s3fs==0.1.8
 fs==2.0.18
 future==0.16.0            # via pyjwkest
 futures==3.2.0 ; python_version == "2.7"
-fuzzywuzzy==0.16.0
-gitdb==0.6.4              # via gitpython
-gitpython==0.3.2.rc1
 glob2==0.3
 gunicorn==0.17.4
 hash-ring==1.3.1          # via django-memcached-hashring
 help-tokens==1.0.3
 html5lib==0.999
 httplib2==0.11.3          # via oauth2, zendesk
-httpretty==0.8.14
 idna==2.6
 ipaddr==2.1.11
 ipaddress==1.0.22
 isodate==0.6.0            # via python-saml
-itsdangerous==0.24        # via flask
-jinja2==2.10              # via flask, moto, sphinx
 jmespath==0.9.3           # via boto3, botocore
 jsondiff==1.1.1           # via edx-enterprise
 jsonfield==2.0.2
@@ -181,25 +157,16 @@ loremipsum==1.0.5
 lxml==3.8.0
 mailsnake==1.6.2
 mako==1.0.2
-mando==0.3.3              # via radon
 markdown==2.6.11
 markey==0.8               # via django-babel-underscore
 markupsafe==1.0
 mock==1.0.1
 mongoengine==0.10.0
-moto==0.3.1
 mysql-python==1.2.5
-needle==0.5.0
 networkx==1.7
 newrelic==3.2.0.91
 nltk==3.2.5
 nodeenv==1.1.1
-nose-exclude==0.5.0
-nose-faulthandler==0.1
-nose-ignore-docstring==0.2
-nose-randomly==1.2.0
-nose-xunitmp==0.3.2
-nose==1.3.7
 numpy==1.6.2
 oauth2==1.9.0.post1
 oauthlib==2.0.1
@@ -208,10 +175,9 @@ pathtools==0.1.2
 paver==1.3.4
 pbr==4.0.2
 pdfminer==20140328
-pep8==1.5.7
 piexif==1.0.2
 pillow==3.4.0
-polib==1.0.3
+polib==1.1.0              # via edx-i18n-tools
 psutil==1.2.1
 py2neo==3.1.2
 pycontracts==1.7.1
@@ -225,61 +191,43 @@ pyjwt==1.5.2
 pymongo==2.9.1
 pynliner==0.5.2
 pyparsing==2.2.0
-pyquery==1.2.9
 pysrt==0.4.7
 python-dateutil==2.4.0
 python-levenshtein==0.12.0
 python-memcached==1.48
-python-mimeparse==1.6.0   # via testtools
 python-openid==2.2.5
 python-saml==2.4.0
-python-subunit==0.0.16
 python-swiftclient==3.5.0
 pytz==2016.10
 pyuca==1.1
 pyyaml==3.12
-radon==1.3.2
 redis==2.10.6
 reportlab==3.1.44
 requests-oauthlib==0.6.1
 requests==2.9.1
 rules==1.3
-s3transfer==0.1.12
+s3transfer==0.1.13        # via boto3
 sailthru-client==2.2.3
 scipy==0.14.0
-selenium==3.11.0
 shapely==1.2.16
 shortuuid==0.5.0          # via edx-django-oauth2-provider
 simplejson==3.13.2        # via dogapi, mailsnake, sailthru-client, zendesk
-singledispatch==3.4.0.2
 six==1.11.0
 slumber==0.7.1            # via edx-rest-api-client
-smmap==0.9.0              # via gitdb
 social-auth-app-django==1.2.0
 social-auth-core==1.4.0
 sorl-thumbnail==12.3
 sortedcontainers==0.9.2
-sphinx==1.1.3
-sphinx_rtd_theme==0.1.5
-splinter==0.5.4
-sqlparse==0.2.4           # via django-debug-toolbar
 stevedore==1.10.0
-sure==1.2.3
 sympy==0.7.1
-testfixtures==4.5.0
-testtools==0.9.34
-text-unidecode==1.2       # via faker
-transifex-client==0.12.1
 unicodecsv==0.14.1
-urllib3==1.22             # via elasticsearch, transifex-client
+urllib3==1.22             # via elasticsearch
 user-util==0.1.3
 voluptuous==0.11.1
 watchdog==0.8.3
 web-fragments==0.2.2
 webob==1.8.1              # via xblock
-werkzeug==0.14.1          # via flask
 wrapt==1.10.5
 xblock-review==1.1.5
 xblock==1.1.1
-xmltodict==0.4.1
 zendesk==1.1.1
