+++ requirements/edx/development.txt
@@ -45,6 +45,7 @@ git+https://github.com/edx-solutions/xblock-drag-and-drop-v2@2.1.5#egg=xblock-dr
 git+https://github.com/open-craft/xblock-poll@7ba819b968fe8faddb78bb22e1fe7637005eb414#egg=xblock-poll==1.2.7
 git+https://github.com/edx/xblock-utils.git@v1.0.5#egg=xblock-utils==1.0.5
 -e common/lib/xmodule
+alabaster==0.7.10         # via sphinx
 amqp==1.4.9
 analytics-python==1.1.0
 anyjson==0.3.3
@@ -69,7 +70,6 @@ botocore==1.8.17
 celery==3.1.25
 cffi==1.11.5
 charade==1.0.3
-chrono==1.0.2
 click-log==0.1.8
 click==6.7
 colorama==0.3.9
@@ -119,7 +119,6 @@ django-user-tasks==0.1.5
 django-waffle==0.12.0
 django-webpack-loader==0.4.1
 django==1.11.12
-django_nose==1.4.1
 djangorestframework-jwt==1.11.0
 djangorestframework-xml==1.3.0
 dm.xmlsec.binding==1.3.3
@@ -135,7 +134,7 @@ edx-django-oauth2-provider==1.2.5
 edx-django-release-util==0.3.1
 edx-django-sites-extensions==2.3.1
 edx-drf-extensions==1.2.5
-edx-enterprise==0.67.4
+edx-enterprise==0.67.5
 edx-i18n-tools==0.4.4
 edx-lint==0.5.4
 edx-milestones==0.1.13
@@ -145,6 +144,7 @@ edx-organizations==0.4.10
 edx-proctoring==1.3.9
 edx-rest-api-client==1.7.1
 edx-search==1.1.0
+edx-sphinx-theme==1.3.0
 edx-submissions==2.0.12
 edx-user-state-client==1.0.4
 edxval==0.1.13
@@ -155,11 +155,9 @@ execnet==1.5.0
 extras==1.0.0
 factory_boy==2.8.1
 faker==0.8.13
-faulthandler==3.0
 feedparser==5.1.3
 firebase-token-generator==1.3.2
 first==2.0.1
-flaky==3.4.0
 flask==0.12.2
 freezegun==0.3.8
 fs-s3fs==0.1.8
@@ -167,8 +165,6 @@ fs==2.0.18
 future==0.16.0
 futures==3.2.0 ; python_version == "2.7"
 fuzzywuzzy==0.16.0
-gitdb==0.6.4
-gitpython==0.3.2.rc1
 glob2==0.3
 gunicorn==0.17.4
 hash-ring==1.3.1
@@ -177,6 +173,7 @@ html5lib==0.999
 httplib2==0.11.3
 httpretty==0.8.14
 idna==2.6
+imagesize==1.0.0          # via sphinx
 incremental==17.5.0
 inflect==0.2.5
 ipaddr==2.1.11
@@ -212,16 +209,12 @@ networkx==1.7
 newrelic==3.2.0.91
 nltk==3.2.5
 nodeenv==1.1.1
-nose-exclude==0.5.0
-nose-faulthandler==0.1
-nose-ignore-docstring==0.2
-nose-randomly==1.2.0
-nose-xunitmp==0.3.2
 nose==1.3.7
 numpy==1.6.2
 oauth2==1.9.0.post1
 oauthlib==2.0.1
 pa11ycrawler==1.6.2
+packaging==17.1           # via sphinx
 parsel==1.4.0
 path.py==8.2.1
 pathtools==0.1.2
@@ -231,9 +224,9 @@ pdfminer==20140328
 pep8==1.5.7
 piexif==1.0.2
 pillow==3.4.0
-pip-tools==1.11.0
+pip-tools==2.0.1
 pluggy==0.6.0
-polib==1.0.3
+polib==1.1.0
 psutil==1.2.1
 py2neo==3.1.2
 py==1.5.3
@@ -286,7 +279,7 @@ reportlab==3.1.44
 requests-oauthlib==0.6.1
 requests==2.9.1
 rules==1.3
-s3transfer==0.1.12
+s3transfer==0.1.13
 sailthru-client==2.2.3
 scipy==0.14.0
 scrapy==1.1.2
@@ -295,18 +288,18 @@ service-identity==17.0.0
 shapely==1.2.16
 shortuuid==0.5.0
 simplejson==3.13.2
-singledispatch==3.4.0.2
+singledispatch==3.4.0.3
 six==1.11.0
 slumber==0.7.1
-smmap==0.9.0
+snowballstemmer==1.2.1    # via sphinx
 social-auth-app-django==1.2.0
 social-auth-core==1.4.0
 sorl-thumbnail==12.3
 sortedcontainers==0.9.2
-sphinx==1.1.3
-sphinx_rtd_theme==0.1.5
+sphinx==1.7.2
+sphinxcontrib-websupport==1.0.1  # via sphinx
 splinter==0.5.4
-sqlparse==0.2.4
+sqlparse==0.2.4           # via django-debug-toolbar
 stevedore==1.10.0
 sure==1.2.3
 sympy==0.7.1
@@ -317,6 +310,7 @@ tox-battery==0.5.1
 tox==2.8.2
 transifex-client==0.12.1
 twisted==16.6.0
+typing==3.6.4             # via sphinx
 unicodecsv==0.14.1
 urllib3==1.22
 urlobject==2.4.3
