+++ requirements/edx/pip-tools.txt
@@ -7,5 +7,5 @@
 
 click==6.7                # via pip-tools
 first==2.0.1              # via pip-tools
-pip-tools==1.11.0
+pip-tools==2.0.1
 six==1.11.0               # via pip-tools
