+++ requirements/edx/testing.txt
@@ -13,7 +13,6 @@
 git+https://github.com/solashirai/crowdsourcehinter.git@518605f0a95190949fe77bd39158450639e2e1dc#egg=crowdsourcehinter-xblock==0.1
 git+https://github.com/mitodl/django-cas.git@afac57bc523f145ae826f4ed3d4fa8b2c86c5364#egg=django-cas==2.1.1
 git+https://github.com/edx/django-celery.git@756cb57aad765cb2b0d37372c1855b8f5f37e6b0#egg=django-celery==3.2.1+edx.2
-git+https://github.com/hmarr/django-debug-toolbar-mongo.git@b0686a76f1ce3532088c4aee6e76b9abe61cc808#egg=django-debug-toolbar-mongo==0.1.10
 git+https://github.com/edx/django-oauth-plus.git@01ec2a161dfc3465f9d35b9211ae790177418316#egg=django-oauth-plus==2.2.9.edx-1
 git+https://github.com/edx/django-openid-auth.git@0.15.1#egg=django-openid-auth==0.15.1
 git+https://github.com/jazzband/django-pipeline.git@d068a019169c9de5ee20ece041a6dea236422852#egg=django-pipeline==1.5.3
@@ -68,10 +67,9 @@ botocore==1.8.17
 celery==3.1.25
 cffi==1.11.5
 charade==1.0.3
-chrono==1.0.2
 click-log==0.1.8          # via edx-lint
 click==6.7
-colorama==0.3.9
+colorama==0.3.9           # via radon
 configparser==3.5.0       # via pylint
 constantly==15.1.0        # via twisted
 coverage==4.2
@@ -81,7 +79,7 @@ cssutils==1.0.2
 ddt==0.8.0
 decorator==4.3.0
 defusedxml==0.4.1
-dicttoxml==1.7.4
+dicttoxml==1.7.4          # via moto
 diff-cover==0.9.8
 django-appconf==1.0.2
 django-babel-underscore==0.5.2
@@ -93,7 +91,6 @@ django-config-models==0.1.8
 django-cors-headers==2.1.0
 django-countries==4.6.1
 django-crum==0.7.2
-django-debug-toolbar==1.8
 django-fernet-fields==0.5
 django-filter==1.0.4
 django-ipware==1.1.0
@@ -117,7 +114,6 @@ django-storages==1.4.1
 django-user-tasks==0.1.5
 django-waffle==0.12.0
 django-webpack-loader==0.4.1
-django_nose==1.4.1
 djangorestframework-jwt==1.11.0
 djangorestframework-xml==1.3.0
 dm.xmlsec.binding==1.3.3
@@ -133,7 +129,7 @@ edx-django-oauth2-provider==1.2.5
 edx-django-release-util==0.3.1
 edx-django-sites-extensions==2.3.1
 edx-drf-extensions==1.2.5
-edx-enterprise==0.67.4
+edx-enterprise==0.67.5
 edx-i18n-tools==0.4.4
 edx-lint==0.5.4
 edx-milestones==0.1.13
@@ -150,22 +146,18 @@ elasticsearch==1.9.0
 enum34==1.1.6
 event-tracking==0.2.4
 execnet==1.5.0            # via pytest-xdist
-extras==1.0.0
+extras==1.0.0             # via python-subunit, testtools
 factory_boy==2.8.1
-faker==0.8.13
-faulthandler==3.0
+faker==0.8.13             # via factory-boy
 feedparser==5.1.3
 firebase-token-generator==1.3.2
-flaky==3.4.0
-flask==0.12.2
+flask==0.12.2             # via moto
 freezegun==0.3.8
 fs-s3fs==0.1.8
 fs==2.0.18
 future==0.16.0
 futures==3.2.0 ; python_version == "2.7"
 fuzzywuzzy==0.16.0
-gitdb==0.6.4
-gitpython==0.3.2.rc1
 glob2==0.3
 gunicorn==0.17.4
 hash-ring==1.3.1
@@ -180,7 +172,7 @@ ipaddr==2.1.11
 ipaddress==1.0.22
 isodate==0.6.0
 isort==4.2.5
-itsdangerous==0.24
+itsdangerous==0.24        # via flask
 jinja2-pluralize==0.3.0
 jinja2==2.10
 jmespath==0.9.3
@@ -195,7 +187,7 @@ loremipsum==1.0.5
 lxml==3.8.0
 mailsnake==1.6.2
 mako==1.0.2
-mando==0.3.3
+mando==0.3.3              # via radon
 markdown==2.6.11
 markey==0.8
 markupsafe==1.0
@@ -204,16 +196,11 @@ mock==1.0.1
 mongoengine==0.10.0
 moto==0.3.1
 mysql-python==1.2.5
-needle==0.5.0
+needle==0.5.0             # via bok-choy
 networkx==1.7
 newrelic==3.2.0.91
 nltk==3.2.5
 nodeenv==1.1.1
-nose-exclude==0.5.0
-nose-faulthandler==0.1
-nose-ignore-docstring==0.2
-nose-randomly==1.2.0
-nose-xunitmp==0.3.2
 nose==1.3.7
 numpy==1.6.2
 oauth2==1.9.0.post1
@@ -229,7 +216,7 @@ pep8==1.5.7
 piexif==1.0.2
 pillow==3.4.0
 pluggy==0.6.0             # via tox
-polib==1.0.3
+polib==1.1.0
 psutil==1.2.1
 py2neo==3.1.2
 py==1.5.3                 # via pytest, pytest-catchlog, tox
@@ -266,7 +253,7 @@ pytest==3.1.3
 python-dateutil==2.4.0
 python-levenshtein==0.12.0
 python-memcached==1.48
-python-mimeparse==1.6.0
+python-mimeparse==1.6.0   # via testtools
 python-openid==2.2.5
 python-saml==2.4.0
 python-subunit==0.0.16
@@ -281,7 +268,7 @@ reportlab==3.1.44
 requests-oauthlib==0.6.1
 requests==2.9.1
 rules==1.3
-s3transfer==0.1.12
+s3transfer==0.1.13
 sailthru-client==2.2.3
 scipy==0.14.0
 scrapy==1.1.2             # via pa11ycrawler
@@ -290,24 +277,20 @@ service-identity==17.0.0  # via scrapy
 shapely==1.2.16
 shortuuid==0.5.0
 simplejson==3.13.2
-singledispatch==3.4.0.2
+singledispatch==3.4.0.3
 six==1.11.0
 slumber==0.7.1
-smmap==0.9.0
 social-auth-app-django==1.2.0
 social-auth-core==1.4.0
 sorl-thumbnail==12.3
 sortedcontainers==0.9.2
-sphinx==1.1.3
-sphinx_rtd_theme==0.1.5
 splinter==0.5.4
-sqlparse==0.2.4
 stevedore==1.10.0
 sure==1.2.3
 sympy==0.7.1
 testfixtures==4.5.0
 testtools==0.9.34
-text-unidecode==1.2
+text-unidecode==1.2       # via faker
 tox-battery==0.5.1
 tox==2.8.2
 transifex-client==0.12.1
@@ -322,7 +305,7 @@ w3lib==1.19.0             # via parsel, scrapy
 watchdog==0.8.3
 web-fragments==0.2.2
 webob==1.8.1
-werkzeug==0.14.1
+werkzeug==0.14.1          # via flask
 wrapt==1.10.5
 xblock-review==1.1.5
 xblock==1.1.1
