+++ setup.cfg
@@ -1,19 +1,3 @@
-[nosetests]
-logging-clear-handlers=1
-with-ignore-docstrings=1
-exclude-dir=lms/envs
-            cms/envs
-
-# Without this flag, nose adds /lib directories to the path,
-# which shadows the xblock library (among other things)
-no-path-adjustment=1
-
-process-timeout=300
-
-# Uncomment the following lines to open pdb when a test fails
-#nocapture=1
-#pdb=1
-
 [tool:pytest]
 DJANGO_SETTINGS_MODULE = lms.envs.test
 addopts = --nomigrations --reuse-db --durations=20
